infrastructure now compiles on my system using nao toolchain and locally installed libs but: i had to add a bunch of static_casts

+++ Framework/NaoTH-Commons/Source/Representations/Infrastructure/Image.h
@@ -274,7 +274,7 @@ namespace naoth
 
     inline unsigned int getIndexSize() const
     {
-      return cameraInfo.getSize();
+      return static_cast<unsigned int> (cameraInfo.getSize());
     }
 
     inline unsigned int getXOffsetFromIndex(const unsigned int i) const
